+++ UpDownForms/Controllers/QuestionController.cs
@@ -17,7 +17,6 @@ namespace UpDownForms.Controllers
 
         public QuestionController(QuestionService questionService)
         {
-
             _questionService = questionService;
         }
 
@@ -25,58 +24,59 @@ namespace UpDownForms.Controllers
         public async Task<ActionResult<IEnumerable<QuestionDetailsDTO>>> GetQuestions()
         {
             var response = await _questionService.GetQuestions();
-            if (!response.Success)
-            {
-                return BadRequest(response.Message);
-            }
-            return Ok(response.Data);
+
+            return Ok(response);
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<QuestionDTO>> GetQuestionById(int id)
         {
-            var response = await _questionService.GetQuestionById(id);
-            if (!response.Success)
+            if (!ModelState.IsValid)
             {
-                return BadRequest(response.Message);
+                return BadRequest();
             }
-            return Ok(response.Data);
+            var response = await _questionService.GetQuestionById(id);
+
+            return Ok(response);
         }
 
         [Authorize]
         [HttpPost]
         public async Task<ActionResult<QuestionDetailsDTO>> PostQuestion([FromBody] CreateQuestionDTO createQuestionDTO)
         {
-            var response = await _questionService.PostQuestion(createQuestionDTO);
-            if (!response.Success)
+            if (!ModelState.IsValid)
             {
-                return BadRequest(response.Message);
+                return BadRequest();
             }
-            return Ok(response.Data);
+            var response = await _questionService.PostQuestion(createQuestionDTO);
+
+            return Ok(response);
         }
 
         [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult<QuestionDTO>> PutQuestion(int id, [FromBody] UpdateQuestionDTO updateQuestionDTO)
         {
-            var response = await _questionService.PutQuestion(id, updateQuestionDTO);
-            if (!response.Success)
+            if (!ModelState.IsValid)
             {
-                return BadRequest(response.Message);
+                return BadRequest();
             }
-            return Ok(response.Data);
+            var response = await _questionService.PutQuestion(id, updateQuestionDTO);
+
+            return Ok(response);
         }
 
         [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult<QuestionDTO>> DeleteQuestion(int id)
         {
-            var response = await _questionService.DeleteQuestion(id);
-            if (!response.Success)
+            if (!ModelState.IsValid)
             {
-                return BadRequest(response.Message);
+                return BadRequest();
             }
-            return Ok(response.Data);
+            var response = await _questionService.DeleteQuestion(id);
+            return Ok(response);
+        
         }
 
 
@@ -84,38 +84,38 @@ namespace UpDownForms.Controllers
         [HttpGet("{questionId}/options")]
         public async Task<ActionResult<IEnumerable<OptionDTO>>> GetOptions(int questionId)
         {
-            var response = await _questionService.GetOptionsByQuestion(questionId);
-            if (!response.Success)
+            if (!ModelState.IsValid)
             {
-                return BadRequest(response.Message);
+                return BadRequest();
             }
-            return Ok(response.Data);
+            var response = await _questionService.GetOptionsByQuestion(questionId);
+            return Ok(response);
         }
 
         [Authorize]
         [HttpPost("{questionId}/options")]
         public async Task<IActionResult> AddOption(int questionId, [FromBody] CreateOptionDTO createOptionDTO)
         {
-            var response = await _questionService.AddOption(questionId, createOptionDTO);
-            if (!response.Success)
+            if (!ModelState.IsValid)
             {
-                return BadRequest(response.Message);
+                return BadRequest();
             }
-            return Ok(response.Data);
-           
+            var response = await _questionService.AddOption(questionId, createOptionDTO);
+            return Ok(response);
         }
 
         [Authorize]
         [HttpDelete("{questionId}/options/{optionId}")]
         public async Task<IActionResult> DeleteOption(int questionId, int optionId)
         {
-            var response = await _questionService.DeleteOption(questionId, optionId);
-            if (!response.Success)
+            if (!ModelState.IsValid)
             {
-                return BadRequest(response.Message);
+                return BadRequest();
             }
-            return Ok(response.Data);
+            var response = await _questionService.DeleteOption(questionId, optionId);
+            return Ok(response);
         }
+    
 
         // implement update?
     }
